refactor: HttpCookie -> RequestHeader에서 세션 생성하도록 변경
