+++ tomcat/src/main/java/org/apache/coyote/http11/HttpCookie.java
@@ -2,15 +2,9 @@ package org.apache.coyote.http11;
 
 import java.util.HashMap;
 import java.util.Map;
-import java.util.UUID;
 import java.util.stream.Collectors;
-import org.apache.catalina.Session;
-import org.apache.catalina.SessionManager;
 
 public class HttpCookie {
-
-    private static final SessionManager sessionManager = SessionManager.getInstance();
-    private static final String JSESSIONID = "JSESSIONID";
 
     private final Map<String, String> values;
 
@@ -35,28 +29,8 @@ public class HttpCookie {
         values.put(key, value);
     }
 
-    public Session getSession(boolean isCreate) {
-        if (has(JSESSIONID)) {
-            String sessionId = getValue(JSESSIONID);
-            return sessionManager.findSession(sessionId);
-        }
-        if (isCreate) {
-            Session session = new Session(UUID.randomUUID().toString());
-            sessionManager.add(session);
-            return session;
-        }
-        return null;
-    }
-
     public String getValue(String key) {
-        if (has(key)) {
-            return values.get(key);
-        }
-        throw new IllegalArgumentException("쿠키가 존재하지 않습니다.");
-    }
-
-    public boolean has(String key) {
-        return values.containsKey(key);
+        return values.getOrDefault(key, null);
     }
 
     public boolean isEmpty() {
